Initial template and class file
This commit contains an nfr.cls class file for NFR applications, as well
as the content of the official file "Template for project description -
Researcher project for experienced scientists and early career
scientists.docx", converted to LaTeX by pandoc. I modified the "Template"
file, to use the nfr document class, so that section A of that document
is the actual structure, using the \section and \subsection commands,
and that the other sections are treated as an appendix.

TODO: The formatting of the appendix is appaling
TODO: The reference list should be in a 9pt font