Update bibliography example
I updated the bibliography example in the "Template" file to use two
columns for the bibliography. Does this save space? I don't know, but I
tend to think so!
Update bibliography example

I updated the bibliography example in the "Template" file to use two
columns for the bibliography. Does this save space? I don't know, but I
tend to think so!