Add a CV template
I wanted to use bibtex for the references in my CV! Added two class
options to nfr.cls (pd vs. cv and manager vs. partner) to get this
working. It's not very sophisticated and uses a lot of complex table
formatting, but a LaTeX user should be able to fill in the template
without problems.